+++ webview-ui/src/components/settings/ApiOptions.tsx
@@ -55,6 +55,7 @@ import { DoubaoProvider } from "./providers/DoubaoProvider"
 import { VertexProvider } from "./providers/VertexProvider"
 import GeminiCliProvider from "./providers/GeminiCliProvider"
 import { RequestyProvider } from "./providers/RequestyProvider"
+import { FireworksProvider } from "./providers/FireworksProvider"
 
 interface ApiOptionsProps {
 	showModelOptions: boolean
@@ -768,95 +769,13 @@ const ApiOptions = ({
 				/>
 			)}
 
-			{selectedProvider === "fireworks" && (
-				<div>
-					<VSCodeTextField
-						value={apiConfiguration?.fireworksApiKey || ""}
-						style={{ width: "100%" }}
-						type="password"
-						onInput={handleInputChange("fireworksApiKey")}
-						placeholder="Enter API Key...">
-						<span style={{ fontWeight: 500 }}>Fireworks API Key</span>
-					</VSCodeTextField>
-					<p
-						style={{
-							fontSize: "12px",
-							marginTop: 3,
-							color: "var(--vscode-descriptionForeground)",
-						}}>
-						This key is stored locally and only used to make API requests from this extension.
-						{!apiConfiguration?.fireworksApiKey && (
-							<VSCodeLink
-								href="https://fireworks.ai/settings/users/api-keys"
-								style={{
-									display: "inline",
-									fontSize: "inherit",
-								}}>
-								You can get a Fireworks API key by signing up here.
-							</VSCodeLink>
-						)}
-					</p>
-					<VSCodeTextField
-						value={apiConfiguration?.fireworksModelId || ""}
-						style={{ width: "100%" }}
-						onInput={handleInputChange("fireworksModelId")}
-						placeholder={"Enter Model ID..."}>
-						<span style={{ fontWeight: 500 }}>Model ID</span>
-					</VSCodeTextField>
-					<p
-						style={{
-							fontSize: "12px",
-							marginTop: 3,
-							color: "var(--vscode-descriptionForeground)",
-						}}>
-						<span style={{ color: "var(--vscode-errorForeground)" }}>
-							(<span style={{ fontWeight: 500 }}>Note:</span> Cline uses complex prompts and works best with Claude
-							models. Less capable models may not work as expected.)
-						</span>
-					</p>
-					<VSCodeTextField
-						value={apiConfiguration?.fireworksModelMaxCompletionTokens?.toString() || ""}
-						style={{ width: "100%", marginBottom: 8 }}
-						onInput={(e) => {
-							const value = (e.target as HTMLInputElement).value
-							if (!value) {
-								return
-							}
-							const num = parseInt(value, 10)
-							if (isNaN(num)) {
-								return
-							}
-							handleInputChange("fireworksModelMaxCompletionTokens")({
-								target: {
-									value: num,
-								},
-							})
-						}}
-						placeholder={"2000"}>
-						<span style={{ fontWeight: 500 }}>Max Completion Tokens</span>
-					</VSCodeTextField>
-					<VSCodeTextField
-						value={apiConfiguration?.fireworksModelMaxTokens?.toString() || ""}
-						style={{ width: "100%", marginBottom: 8 }}
-						onInput={(e) => {
-							const value = (e.target as HTMLInputElement).value
-							if (!value) {
-								return
-							}
-							const num = parseInt(value)
-							if (isNaN(num)) {
-								return
-							}
-							handleInputChange("fireworksModelMaxTokens")({
-								target: {
-									value: num,
-								},
-							})
-						}}
-						placeholder={"4000"}>
-						<span style={{ fontWeight: 500 }}>Max Context Tokens</span>
-					</VSCodeTextField>
-				</div>
+			{apiConfiguration && selectedProvider === "fireworks" && (
+				<FireworksProvider
+					apiConfiguration={apiConfiguration}
+					handleInputChange={handleInputChange}
+					showModelOptions={showModelOptions}
+					isPopup={isPopup}
+				/>
 			)}
 
 			{selectedProvider === "vscode-lm" && (
@@ -1527,6 +1446,7 @@ const ApiOptions = ({
 				selectedProvider !== "doubao" &&
 				selectedProvider !== "vertex" &&
 				selectedProvider !== "gemini-cli" &&
+				selectedProvider !== "fireworks" &&
 				showModelOptions && (
 					<>
 						<DropdownContainer zIndex={DROPDOWN_Z_INDEX - 2} className="dropdown-container">
